add number format on numbers and change print design

+++ database/migrations/2023_11_23_073352_create_invoice_totals.php
@@ -14,8 +14,8 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('invoice_id')->constrained('invoices');
             $table->decimal('discount')->nullable();
-            $table->decimal('vat')->nullable();
-            $table->decimal('grand_price');
+            $table->double('vat')->nullable();
+            $table->double('grand_price');
             $table->timestamps();
         });
     }
